+++ jre_emul/android/libcore/luni/src/main/java/java/util/zip/Deflater.java
@@ -281,8 +281,8 @@ public class Deflater {
           @throw AUTORELEASE([[JavaUtilZipDataFormatException alloc] init]);
       }
 
-      int bytesRead = zStream->next_in - initialNextIn;
-      int bytesWritten = zStream->next_out - initialNextOut;
+      int bytesRead = (int) (zStream->next_in - initialNextIn);
+      int bytesWritten = (int) (zStream->next_out - initialNextOut);
 
       self->inRead_ += bytesRead;
       return bytesWritten;
@@ -361,7 +361,7 @@ public class Deflater {
 
     private native int getAdlerImpl(long handle) /*-[
         z_stream *zStream = (z_stream*) handle;
-        return zStream->adler;
+        return (int) zStream->adler;
     ]-*/;
 
     /**
